drift from float to int

+++ firmware/src/WebServer.cpp
@@ -31,7 +31,7 @@ void web_server_init()
 		if (request->hasParam(HTTP_CONFIG) && request->hasParam(HTTP_SPEED)) {
 			robot_save_configuration(request->getParam(HTTP_CONFIG)->value().toInt(),
 			                         request->getParam(HTTP_SPEED)->value().toInt(),
-			                         request->getParam(HTTP_MOTOR_DRIFT)->value().toFloat());
+			                         request->getParam(HTTP_MOTOR_DRIFT)->value().toInt());
 		}
 		else {
 			robot_connect();
